+++ lambdas/sqs-message-remover/variables.tf
@@ -6,7 +6,7 @@ variable "prefix" {
 }
 
 variable "cmr_environment" {
-  description = "Environment that should be used for CMR requests (e.g. 'UAT', 'SIT')"
+  description = "Custom host or logical environment (e.g. 'UAT', 'SIT') that should be used for CMR requests"
   type        = string
 }