+++ tf-modules/cumulus/variables.tf
@@ -12,7 +12,7 @@ variable "cmr_client_id" {
 }
 
 variable "cmr_environment" {
-  description = "Environment that should be used for CMR requests (e.g. 'UAT', 'SIT')"
+  description = "Custom host or logical environment (e.g. 'UAT', 'SIT') that should be used for CMR requests"
   type        = string
 }